新增重新开始按钮

+++ src/edu/hitsz/aircraft/AbstractAircraft.java
@@ -15,16 +15,16 @@ public abstract class AbstractAircraft extends AbstractFlyingObject {
     /**
      * 生命值
      */
-    protected int maxHp;
-    protected int hp;
+    protected double maxHp;
+    protected double hp;
 
-    public AbstractAircraft(int locationX, int locationY, int speedX, double speedY, int hp) {
+    public AbstractAircraft(int locationX, int locationY, int speedX, double speedY, double hp) {
         super(locationX, locationY, speedX, speedY);
         this.hp = hp;
         this.maxHp = hp;
     }
 
-    public void decreaseHp(int decrease){
+    public void decreaseHp(double decrease){
         hp -= decrease;
         if(hp <= 0){
             hp=0;
@@ -32,7 +32,7 @@ public abstract class AbstractAircraft extends AbstractFlyingObject {
         }
     }
 
-    public int getHp() {
+    public double getHp() {
         return hp;
     }
